@@ -83,40 +83,6 @@ func (my *Manager) goWatchLocalExcel(later loom.Later) {
 	}
 }
 
-// 本来使用 github.com/fsnotify/fsnotify ，测试了三次，结果分别是：op=Remove, op=Chmod, 以及没有收到event
-//func (my *Manager) goWatchLocalExcel(rawFilePath string, args ExcelArgs) {
-//	watcher, err := fsnotify.NewWatcher()
-//	if err != nil {
-//		logo.JsonE("err", err)
-//	}
-//	defer watcher.Close()
-//
-//	err = watcher.Add(rawFilePath)
-//	if err != nil {
-//		logo.JsonE("err", err)
-//	}
-//
-//	for {
-//		select {
-//		case event, ok := <-watcher.Events:
-//			if !ok {
-//				return
-//			}
-//
-//			if event.Op&fsnotify.Write == fsnotify.Write {
-//				my.addLocalExcel(rawFilePath, args)
-//			}
-//			logo.JsonI("event", event)
-//		case err, ok := <-watcher.Errors:
-//			if !ok {
-//				return
-//			}
-//
-//			logo.JsonW("err", err)
-//		}
-//	}
-//}
-
 func (my *Manager) addLocalExcel(rawFilePath string, args ExcelArgs) {
 	var sheetNames = loadSheetNames(args.FilePath)
 	for _, name := range sheetNames {
@@ -135,20 +101,20 @@ func (my *Manager) OnExcelChanged(handler func(excelFilePath string)) {
 	my.onExcelChanged.Add(handler)
 }
 
-func (my *Manager) GetTemplate(pTemplate interface{}, id interface{}, opts ...Option) bool {
+func (my *Manager) GetTemplate(pTemplate any, id any, opts ...Option) bool {
 	var args = my.createOptions(opts)
 	var manager = (*TemplateManager)(atomic.LoadPointer(&my.templateManager))
 	return manager != nil && id != nil && manager.getTemplate(&my.routeTable, pTemplate, id, args.SheetName)
 }
 
-// 相同的参数每次返回的pTemplateList中的items的不保证顺序：这个是跟实现相关的，目前遍历基于map是不稳定的
-func (my *Manager) GetTemplates(pTemplateList interface{}, opts ...Option) bool {
+// GetTemplates 相同的参数每次返回的pTemplateList中的items的不保证顺序：这个是跟实现相关的，目前遍历基于map是不稳定的
+func (my *Manager) GetTemplates(pTemplateList any, opts ...Option) bool {
 	var args = my.createOptions(opts)
 	var manager = (*TemplateManager)(atomic.LoadPointer(&my.templateManager))
 	return manager != nil && manager.getTemplates(&my.routeTable, pTemplateList, args)
 }
 
-func (my *Manager) GetConfig(pConfig interface{}, opts ...Option) bool {
+func (my *Manager) GetConfig(pConfig any, opts ...Option) bool {
 	var args = my.createOptions(opts)
 	var manager = (*ConfigManager)(atomic.LoadPointer(&my.configManager))
 	return manager != nil && manager.getConfig(&my.routeTable, pConfig, args.SheetName)
@@ -183,3 +149,37 @@ func (my *Manager) isSheetName(name string) bool {
 	var _, ok = my.routeTable.Load(name)
 	return ok
 }
+
+// 本来使用 github.com/fsnotify/fsnotify ，测试了三次，结果分别是：op=Remove, op=Chmod, 以及没有收到event
+//func (my *Manager) goWatchLocalExcel(rawFilePath string, args ExcelArgs) {
+//	watcher, err := fsnotify.NewWatcher()
+//	if err != nil {
+//		logo.JsonE("err", err)
+//	}
+//	defer watcher.Close()
+//
+//	err = watcher.Add(rawFilePath)
+//	if err != nil {
+//		logo.JsonE("err", err)
+//	}
+//
+//	for {
+//		select {
+//		case event, ok := <-watcher.Events:
+//			if !ok {
+//				return
+//			}
+//
+//			if event.Op&fsnotify.Write == fsnotify.Write {
+//				my.addLocalExcel(rawFilePath, args)
+//			}
+//			logo.JsonI("event", event)
+//		case err, ok := <-watcher.Errors:
+//			if !ok {
+//				return
+//			}
+//
+//			logo.JsonW("err", err)
+//		}
+//	}
+//}
